Test fixes and tweaks (#20870)
Issues were exposed by running the tests under
the Rust options parser.
