+++ src/python/pants/bin/remote_pants_runner.py
@@ -19,6 +19,7 @@ from pants.engine.internals.native_engine import (
 from pants.option.global_options import GlobalOptions
 from pants.option.options_bootstrapper import OptionsBootstrapper
 from pants.pantsd.pants_daemon_client import PantsDaemonClient
+from pants.util.strutil import get_strict_env
 
 logger = logging.getLogger(__name__)
 
@@ -106,7 +107,7 @@ class RemotePantsRunner:
         :param options_bootstrapper: The bootstrap options.
         """
         self._args = args
-        self._env = env
+        self._env = get_strict_env(env, logger)
         self._options_bootstrapper = options_bootstrapper
         self._bootstrap_options = options_bootstrapper.bootstrap_options
         self._client = PantsDaemonClient(self._bootstrap_options)
@@ -130,37 +131,16 @@ class RemotePantsRunner:
         executor: PyExecutor,
         start_time: float,
     ) -> ExitCode:
-        global_options = self._bootstrap_options.for_global_scope()
-
-        # NB: If an OS string is not valid UTF-8, Python encodes the non-decodable bytes
-        #  as lone surrogates (see https://peps.python.org/pep-0383/).
-        #  However when we pass these to Rust, we will fail to decode as strict UTF-8.
-        #  So we perform a lossy re-encoding to prevent this.
-        def strict_utf8(s: str) -> str:
-            return s.encode("utf-8", "replace").decode("utf-8")
-
-        strict_env = {}
-        for key, val in sorted(self._env.items()):
-            strict_key = strict_utf8(key)
-            strict_val = strict_utf8(val)
-            if strict_key == key:
-                if strict_val == val:
-                    strict_env[strict_key] = strict_val
-                else:
-                    logger.warning(f"Environment variable with non-UTF-8 value ignored: {key}")
-            else:
-                # We can only log strict_key, because logging will choke on non-UTF-8.
-                # But the reader will know what we mean.
-                logger.warning(f"Environment variable with non-UTF-8 name ignored: {strict_key}")
-
         try:
             # Merge the nailgun TTY capability environment variables with the passed environment dict.
             ng_env = ttynames_to_env(sys.stdin, sys.stdout, sys.stderr)
         except OSError as e:
             logger.debug("Failed to execute ttynames_to_env:  %s", e)
             ng_env = {}
+
+        global_options = self._bootstrap_options.for_global_scope()
         modified_env = {
-            **strict_env,
+            **self._env,
             **ng_env,
             "PANTSD_RUNTRACKER_CLIENT_START_TIME": str(start_time),
             "PANTSD_REQUEST_TIMEOUT_LIMIT": str(
