+++ src/python/pants/util/strutil.py
@@ -10,7 +10,8 @@ import re
 import shlex
 import textwrap
 from collections import abc
-from typing import Any, Callable, Iterable, TypeVar
+from logging import Logger
+from typing import Any, Callable, Iterable, Mapping, TypeVar
 
 import colors
 from typing_extensions import ParamSpec
@@ -379,3 +380,28 @@ def stable_hash(value: Any, *, name: str = "sha256") -> str:
             value, indent=None, separators=(",", ":"), sort_keys=True, cls=_JsonEncoder
         ).encode("utf-8"),
     ).hexdigest()
+
+
+# NB: If an OS string is not valid UTF-8, Python encodes the non-decodable bytes
+#  as lone surrogates (see https://peps.python.org/pep-0383/).
+#  However when we pass these to Rust, we will fail to decode as strict UTF-8.
+#  So we perform a lossy re-encoding to prevent this.
+def strict_utf8(s: str) -> str:
+    return s.encode("utf-8", "replace").decode("utf-8")
+
+
+def get_strict_env(env: Mapping[str, str], logger: Logger) -> Mapping[str, str]:
+    strict_env = {}
+    for key, val in sorted(env.items()):
+        strict_key = strict_utf8(key)
+        strict_val = strict_utf8(val)
+        if strict_key == key:
+            if strict_val == val:
+                strict_env[strict_key] = strict_val
+            else:
+                logger.warning(f"Environment variable with non-UTF-8 value ignored: {key}")
+        else:
+            # We can only log strict_key, because logging will choke on non-UTF-8.
+            # But the reader will know what we mean.
+            logger.warning(f"Environment variable with non-UTF-8 name ignored: {strict_key}")
+    return strict_env
